2011-05-26  Andreas Kling  <kling@webkit.org>
        Reviewed by Benjamin Poulain.

        JSC DOM bindings: ExecState::argumentCount() is size_t, not int.
        https://bugs.webkit.org/show_bug.cgi?id=61530

        * bindings/scripts/CodeGeneratorJS.pm:
        * bindings/scripts/test/JS/JSTestObj.cpp:
        (WebCore::jsTestObjPrototypeFunctionOptionsObject):
        (WebCore::jsTestObjPrototypeFunctionWithDynamicFrameAndOptionalArg):
        (WebCore::jsTestObjPrototypeFunctionWithDynamicFrameAndUserGestureASAD):
        (WebCore::jsTestObjPrototypeFunctionMethodWithOptionalArg):
        (WebCore::jsTestObjPrototypeFunctionMethodWithNonOptionalArgAndOptionalArg):
        (WebCore::jsTestObjPrototypeFunctionMethodWithNonOptionalArgAndTwoOptionalArgs):
        (WebCore::jsTestObjPrototypeFunctionOverloadedMethod2):
        (WebCore::jsTestObjPrototypeFunctionClassMethodWithOptional):

Canonical link: https://commits.webkit.org/76941@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@87390 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/bindings/scripts/test/JS/JSTestObj.cpp
@@ -1094,7 +1094,7 @@ EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionOptionsObject(ExecState* 
     if (exec->hadException())
         return JSValue::encode(jsUndefined());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 1) {
         imp->optionsObject(oo);
         return JSValue::encode(jsUndefined());
@@ -1244,7 +1244,7 @@ EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionWithDynamicFrameAndOption
     if (exec->hadException())
         return JSValue::encode(jsUndefined());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 1) {
         imp->withDynamicFrameAndOptionalArg(dynamicFrame, intArg);
         return JSValue::encode(jsUndefined());
@@ -1292,7 +1292,7 @@ EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionWithDynamicFrameAndUserGe
     if (exec->hadException())
         return JSValue::encode(jsUndefined());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 1) {
         imp->withDynamicFrameAndUserGestureASAD(dynamicFrame, intArg);
         return JSValue::encode(jsUndefined());
@@ -1393,7 +1393,7 @@ EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionMethodWithOptionalArg(Exe
     ASSERT_GC_OBJECT_INHERITS(castedThis, &JSTestObj::s_info);
     TestObj* imp = static_cast<TestObj*>(castedThis->impl());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 0) {
         imp->methodWithOptionalArg();
         return JSValue::encode(jsUndefined());
@@ -1419,7 +1419,7 @@ EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionMethodWithNonOptionalArgA
     if (exec->hadException())
         return JSValue::encode(jsUndefined());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 1) {
         imp->methodWithNonOptionalArgAndOptionalArg(nonOpt);
         return JSValue::encode(jsUndefined());
@@ -1445,7 +1445,7 @@ EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionMethodWithNonOptionalArgA
     if (exec->hadException())
         return JSValue::encode(jsUndefined());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 1) {
         imp->methodWithNonOptionalArgAndTwoOptionalArgs(nonOpt);
         return JSValue::encode(jsUndefined());
@@ -1553,7 +1553,7 @@ static EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionOverloadedMethod2(
     if (exec->hadException())
         return JSValue::encode(jsUndefined());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 1) {
         imp->overloadedMethod(objArg);
         return JSValue::encode(jsUndefined());
@@ -1654,7 +1654,7 @@ EncodedJSValue JSC_HOST_CALL jsTestObjPrototypeFunctionClassMethodWithOptional(E
     ASSERT_GC_OBJECT_INHERITS(castedThis, &JSTestObj::s_info);
     TestObj* imp = static_cast<TestObj*>(castedThis->impl());
 
-    int argsCount = exec->argumentCount();
+    size_t argsCount = exec->argumentCount();
     if (argsCount <= 0) {
 
         JSC::JSValue result = jsNumber(imp->classMethodWithOptional());
